:tada: corrected ortograph faults + Improved the donation banner logic, and finally decided to integrated it with the Footer (when doing a research on the main pahe GPTs, the banner was moving regarding the number of results display) + Change the picture of the Family Flow project for a better visual + added the year of creation of each project present in the potfolio + Added my skills in automation in the Profil component + upgraded the anchors conducing to, or the GPTs page, or to the Portfolio page for a better SEO
