+++ components/Footer/Footer.module.scss
@@ -1,3 +1,8 @@
+// VARIABLES
+@import "variables";
+// MIXINS
+@import "mixins";
+
 // FOOTER SECTION
 .footer {
   width: 100%;
@@ -43,4 +48,145 @@
     margin-right: 2rem;
     font-size: 0.9rem;
   }
+}
+
+// DONATION BANNER
+.donationBanner {
+  width: 100%;
+  height: 4rem;
+  background-color: $navigation-color;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  @media (max-width: 1024px) {
+    width: fit-content;
+    border-radius: 25px;
+    padding: 1.2rem;
+  }
+
+  @media (max-width: 768px) {
+    padding: 1rem;
+  }
+}
+
+.clickHere {
+  font-size: 1.2rem;
+  font-weight: 500;
+  color: white;
+
+  @media (max-width: 1024px) {
+    font-size: 1rem;
+  }
+}
+
+// MODAL DONATION
+
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.6);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.modal {
+  background-color: $gpts-background-page-color;
+  padding: 2rem;
+  border-radius: 8px;
+  border: 1px solid black;
+  width: 30%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  @media (max-width: 1024px) {
+    width: 50%;
+  }
+
+  @media (max-width: 768px) {
+    width: 70%;
+  }
+
+  @media (max-width: 540px) {
+    width: 90%;
+    padding: 1.2rem;
+  }
+}
+
+.titleDonation {
+  color: $navigation-color;
+  font-size: 1.6rem;
+  font-weight: bold;
+  margin-bottom: 1rem;
+
+  @media (max-width: 540px) {
+    font-size: 1.4rem;
+  }
+}
+
+.donationOptions {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  margin: 20px 0 30px 0;
+
+  @media (max-width: 540px) {
+    gap: 10px;
+  }
+}
+
+.donationOption {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+}
+
+.donationText {
+  color: black;
+  font-size: 1.2rem;
+  font-weight: 700;
+  margin-bottom: 0.3rem;
+
+  @media (max-width: 540px) {
+    font-size: 1rem;
+  }
+}
+
+.donationCopy {
+  width: fit-content;
+  display: flex;
+  gap: 0.6rem;
+  background-color: white;
+  padding: 0.5rem 1rem;
+  border: 1px solid black;
+  border-radius: 10px;
+
+  @media (max-width: 540px) {
+    padding: 0.3rem 0.6rem;
+  }
+}
+
+.donationInfos {
+  font-size: 1rem;
+  font-weight: 500;
+  color: black;
+
+  @media (max-width: 540px) {
+    font-size: 0.6rem;
+  }
+}
+
+.copiedNotification {
+  color: green;
+  text-align: center;
+}
+
+.buttonReverted {
+  @include button-style-reverted;
 }
