+++ components/GPTs/GPTs.tsx
@@ -7,7 +7,6 @@ import GPTs_Card from "../GPTs_Card/GPTs_Card";
 import GPTs_Card_Category from "../GPTs_Card_Category/GPTs_Card_Category";
 import GPTs_Card_Type from "../../types/GPTs_Card_Type";
 import GPTS_Card_Category_Type from "../../types/GPTs_Card_Category";
-import { BsCopy } from "react-icons/bs";
 import { CiSearch } from "react-icons/ci";
 import classes from "./GPTs.module.scss";
 
@@ -24,10 +23,6 @@ const GPTs = () => {
   const [searchTerm, setSearchTerm] = useState("");
   //  State to store if the search has been executed
   const [searchExecuted, setSearchExecuted] = useState(false);
-  // State to store if the modal should be shown
-  const [showModal, setShowModal] = useState(false);
-  // State to store the information copied to the clipboard
-  const [copiedInfo, setCopiedInfo] = useState("");
   // Determine the GPTs link based on the active language
   const portfolioLink = activeLanguage === "FR" ? `/fr` : `/`;
 
@@ -102,13 +97,6 @@ const GPTs = () => {
     setSearchTerm(event.target.value);
   };
 
-  // Function to handle the copy to clipboard
-  const handleCopy = (info) => {
-    navigator.clipboard.writeText(info);
-    setCopiedInfo(`Info Copied`);
-    setTimeout(() => setCopiedInfo(""), 2000); // Clear notification after 2 seconds
-  };
-
   //   Object to store the translations for the title and subtitle
   const translations: TranslationsType = {
     title: {
@@ -136,32 +124,20 @@ const GPTs = () => {
       FR: "Collections de GPTs",
     },
     promo: {
-      EN: "Want to know more about The Wise Duck Dev ? Click ",
-      FR: "Vous voulez en savoir plus sur The Wise Duck Dev ? Cliquez ",
+      EN: "Learn more about The Wise Duck Dev ",
+      FR: "Apprenez-en plus sur The Wise Duck Dev ",
     },
     search_aria_label: {
       EN: "Search GPTs on the site",
       FR: "Rechercher des GPTs sur le site",
     },
-    donationBanner: {
-      EN: "To support this project, click here",
-      FR: "Pour soutenir ce projet, cliquez ici",
+    buttonReset: {
+      EN: "Reset",
+      FR: "Effacer",
     },
-    donationTitle: {
-      EN: "Support Our Mission",
-      FR: "Soutenez notre mission",
-    },
-    donationText: {
-      EN: "Your donation enables us to maintain and expand our philanthropic platform, empowering developers with innovative tools for a transformative future. Help us continue to redefine technology, donate today!",
-      FR: "Votre don nous permet de maintenir et d'élargir notre plateforme philanthropique, permettant aux développeurs d'accéder à des outils innovants pour un avenir transformateur. Aidez-nous à continuer à redéfinir la technologie, faites un don aujourd'hui !",
-    },
-    BTC: {
-      EN: "BTC Address: ",
-      FR: "Adresse BTC: ",
-    },
-    ETH: {
-      EN: "ETH Address: ",
-      FR: "Adresse ETH: ",
+    buttonSearch: {
+      EN: "Search",
+      FR: "Chercher",
     },
   };
 
@@ -185,13 +161,12 @@ const GPTs = () => {
       </div>
       <div className={classes.buttonsContainer}>
         <button onClick={resetSearch} className={classes.buttonReverted}>
-          Reset
+          {translations.buttonReset[activeLanguage]}
         </button>
         <button onClick={handleSearchClick} className={classes.button}>
-          Search
+          {translations.buttonSearch[activeLanguage]}
         </button>
       </div>
-
       {searchExecuted ? (
         <div className={classes.resultsInfo}>
           {filteredGPTs.length > 0 ? (
@@ -211,7 +186,6 @@ const GPTs = () => {
           {translations.collection[activeLanguage]}
         </h2>
       )}
-
       <div
         className={
           filteredGPTs.length === 1
@@ -246,87 +220,12 @@ const GPTs = () => {
               />
             ))}
       </div>
-      <p className={classes.promoText}>
+      <Link href={portfolioLink} className={classes.promoText}>
         {translations.promo[activeLanguage]}{" "}
-        <Link href={portfolioLink} className={`${classes.portfolioLinkOrange}`}>
-          {" "}
+        <span className={classes.portfolioLinkOrange}>
           {activeLanguage === "FR" ? "ici" : "here"}
-        </Link>{" "}
-      </p>
-      <div className={classes.donationBanner}>
-        <button
-          onClick={() => setShowModal(true)}
-          className={classes.clickHere}
-        >
-          {translations.donationBanner[activeLanguage]}
-        </button>
-      </div>
-      {showModal && (
-        <div className={classes.overlay}>
-          <div className={classes.modal}>
-            <h2 className={classes.titleDonation}>
-              {translations.donationTitle[activeLanguage]}
-            </h2>
-            <p>{translations.donationText[activeLanguage]}</p>
-
-            <div className={classes.donationOptions}>
-              <div className={classes.donationOption}>
-                <p className={classes.donationText}>Paypal: </p>
-                <div className={classes.donationCopy}>
-                  <p className={classes.donationInfos}>@wiseduckdev</p>
-                  <button onClick={() => handleCopy("@wiseduckdev")}>
-                    <BsCopy />
-                  </button>
-                </div>
-              </div>
-              <div className={classes.donationOption}>
-                <p className={classes.donationText}>
-                  {translations.BTC[activeLanguage]}
-                </p>
-                <div className={classes.donationCopy}>
-                  <p className={classes.donationInfos}>
-                    bc1qjalnp6rhvalxeyclcwfcp0madsmghcfzmylxf4
-                  </p>
-                  <button
-                    onClick={() =>
-                      handleCopy("bc1qjalnp6rhvalxeyclcwfcp0madsmghcfzmylxf4")
-                    }
-                  >
-                    <BsCopy />
-                  </button>
-                </div>
-              </div>
-              <div className={classes.donationOption}>
-                <p className={classes.donationText}>
-                  {translations.ETH[activeLanguage]}
-                </p>
-                <div className={classes.donationCopy}>
-                  <p className={classes.donationInfos}>
-                    0x97553A534aD05fb7D67aFa8fd4d8BCC8C990b477
-                  </p>
-                  <button
-                    onClick={() =>
-                      handleCopy("0x97553A534aD05fb7D67aFa8fd4d8BCC8C990b477")
-                    }
-                  >
-                    <BsCopy />
-                  </button>
-                </div>
-              </div>
-              {copiedInfo && (
-                <p className={classes.copiedNotification}>{copiedInfo}</p>
-              )}
-            </div>
-
-            <button
-              onClick={() => setShowModal(false)}
-              className={classes.buttonReverted}
-            >
-              Close
-            </button>
-          </div>
-        </div>
-      )}
+        </span>
+      </Link>
     </section>
   );
 };
