+++ components/Modals/modalProject/modalProject.module.scss
@@ -112,7 +112,7 @@
   flex-wrap: wrap;
   gap: 1rem;
   padding: 1rem 0;
-  margin: 1rem 0 2rem 0;
+  margin: 1rem 0 1rem 0;
 }
 
 // TECHNOLOGY PROJECT TITLE
@@ -130,7 +130,7 @@
 .buttonContainer {
   display: flex;
   justify-content: center;
-  margin-top: 1rem;
+  margin-top: 2rem;
   gap: 1rem;
 }
