+++ components/Tips/Tips.tsx
@@ -118,8 +118,8 @@ function Tips() {
       FR: "Mise à jour des informations par la recherche en ligne",
     },
     point11: {
-      EN: "In a rapidly evolving field like web, mobile, and blockchain development,and considering that the information used by the GPTs stop in march 2023, access to up-to-date information is indispensable. Explicitly indicate when you need the GPT to perform online searches to ensure that the advice received is based on the latest trends and technologies, thus ensuring that your projects remain at the forefront of innovation.",
-      FR: "Dans un domaine en évolution rapide comme le développement web, mobile et blockchain, et compte tenu du fait que les informations utilisées par les GPTs s'arrêtent en mars 2023, l'accès à des informations à jour est indispensable. Indiquez explicitement quand vous avez besoin que le GPT effectue des recherches en ligne pour vous assurer que les conseils reçus sont basés sur les dernières tendances et technologies, garantissant ainsi que vos projets restent à la pointe de l'innovation.",
+      EN: "In a rapidly evolving field like web, mobile, artificial intelligence, and blockchain development,and considering that the information used by the GPTs stop in march 2023, access to up-to-date information is indispensable. Explicitly indicate when you need the GPT to perform online searches to ensure that the advice received is based on the latest trends and technologies, thus ensuring that your projects remain at the forefront of innovation.",
+      FR: "Dans un domaine en évolution rapide comme le développement web, mobile, d'intelligence artificielle, ou encore de blockchain, et compte tenu du fait que les informations utilisées par les GPTs s'arrêtent en mars 2023, l'accès à des informations à jour est indispensable. C'est pour cette raison qu'il est absolument vital que demandiez clairemenent et explicitement au GPT d'effectuer des recherches en ligne lorsque vous en avez besoin pour vous assurer que les conseils reçus sont bien basés sur les dernières tendances et technologies, garantissant ainsi que vos projets restent à la pointe de l'innovation.",
     },
     button: {
       EN: "All GPTs",
