+++ pages/gpts/[...slug].tsx
@@ -82,10 +82,6 @@ export default function GPTsSlug({ initialPageData }) {
   const siteUrl = process.env.NEXT_PUBLIC_SITE_URL;
   // Determine the GPTs link based on the active language
   const portfolioLink = activeLanguage === "FR" ? `/fr` : `/`;
-  // State to store if the modal should be shown
-  const [showModal, setShowModal] = useState(false);
-  // State to store the information copied to the clipboard
-  const [copiedInfo, setCopiedInfo] = useState("");
 
   // Set loading to false when the component mounts
   useEffect(() => {
@@ -157,13 +153,6 @@ export default function GPTsSlug({ initialPageData }) {
     );
   };
 
-  // Function to handle the copy to clipboard
-  const handleCopy = (info) => {
-    navigator.clipboard.writeText(info);
-    setCopiedInfo(`Info Copied`);
-    setTimeout(() => setCopiedInfo(""), 2000); // Clear notification after 2 seconds
-  };
-
   // Define the translations
   const translation: TranslationsType = {
     button: {
@@ -179,8 +168,8 @@ export default function GPTsSlug({ initialPageData }) {
       FR: "Partager ",
     },
     promo: {
-      EN: "Want to know more about The Wise Duck Dev ? Click ",
-      FR: "Vous voulez en savoir plus sur The Wise Duck Dev ? Cliquez ",
+      EN: "Learn more about The Wise Duck Dev ",
+      FR: "Apprenez-en plus sur The Wise Duck Dev ",
     },
     tips: {
       EN: "Need some prompting tips ?",
@@ -189,26 +178,6 @@ export default function GPTsSlug({ initialPageData }) {
     aria_label_run: {
       EN: "Run the GPT on OpenAI website",
       FR: "Exécuter le GPT sur le site OpenAI",
-    },
-    donationBanner: {
-      EN: "To support this project, click here",
-      FR: "Pour soutenir ce projet, cliquez ici",
-    },
-    donationTitle: {
-      EN: "Support Our Mission",
-      FR: "Soutenez notre mission",
-    },
-    donationText: {
-      EN: "Your donation enables us to maintain and expand our philanthropic platform, empowering developers with innovative tools for a transformative future. Help us continue to redefine technology, donate today!",
-      FR: "Votre don nous permet de maintenir et d'élargir notre plateforme philanthropique, permettant aux développeurs d'accéder à des outils innovants pour un avenir transformateur. Aidez-nous à continuer à redéfinir la technologie, faites un don aujourd'hui !",
-    },
-    BTC: {
-      EN: "BTC Address: ",
-      FR: "Adresse BTC: ",
-    },
-    ETH: {
-      EN: "ETH Address: ",
-      FR: "Adresse ETH: ",
     },
   };
 
@@ -499,94 +468,12 @@ export default function GPTsSlug({ initialPageData }) {
             <div className={classes.buttonContainer}>
               <RedirectToSearchSectionButton />
             </div>
-            <p className={classes.promoText}>
+            <Link href={portfolioLink} className={classes.promoText}>
               {translation.promo[activeLanguage]}{" "}
-              <Link
-                href={portfolioLink}
-                className={`${classes.portfolioLinkOrange}`}
-              >
-                {" "}
+              <span className={classes.portfolioLinkOrange}>
                 {activeLanguage === "FR" ? "ici" : "here"}
-              </Link>{" "}
-            </p>
-            <div className={classes.donationBanner}>
-              <button
-                onClick={() => setShowModal(true)}
-                className={classes.clickHere}
-              >
-                {translation.donationBanner[activeLanguage]}
-              </button>
-            </div>
-            {showModal && (
-              <div className={classes.overlay}>
-                <div className={classes.modal}>
-                  <h2 className={classes.titleDonation}>
-                    {translation.donationTitle[activeLanguage]}
-                  </h2>
-                  <p>{translation.donationText[activeLanguage]}</p>
-
-                  <div className={classes.donationOptions}>
-                    <div className={classes.donationOption}>
-                      <p className={classes.donationText}>Paypal: </p>
-                      <div className={classes.donationCopy}>
-                        <p className={classes.donationInfos}>@wiseduckdev</p>
-                        <button onClick={() => handleCopy("@wiseduckdev")}>
-                          <BsCopy />
-                        </button>
-                      </div>
-                    </div>
-                    <div className={classes.donationOption}>
-                      <p className={classes.donationText}>
-                        {translation.BTC[activeLanguage]}
-                      </p>
-                      <div className={classes.donationCopy}>
-                        <p className={classes.donationInfos}>
-                          bc1qjalnp6rhvalxeyclcwfcp0madsmghcfzmylxf4
-                        </p>
-                        <button
-                          onClick={() =>
-                            handleCopy(
-                              "bc1qjalnp6rhvalxeyclcwfcp0madsmghcfzmylxf4"
-                            )
-                          }
-                        >
-                          <BsCopy />
-                        </button>
-                      </div>
-                    </div>
-                    <div className={classes.donationOption}>
-                      <p className={classes.donationText}>
-                        {translation.ETH[activeLanguage]}
-                      </p>
-                      <div className={classes.donationCopy}>
-                        <p className={classes.donationInfos}>
-                          0x97553A534aD05fb7D67aFa8fd4d8BCC8C990b477
-                        </p>
-                        <button
-                          onClick={() =>
-                            handleCopy(
-                              "0x97553A534aD05fb7D67aFa8fd4d8BCC8C990b477"
-                            )
-                          }
-                        >
-                          <BsCopy />
-                        </button>
-                      </div>
-                    </div>
-                    {copiedInfo && (
-                      <p className={classes.copiedNotification}>{copiedInfo}</p>
-                    )}
-                  </div>
-
-                  <button
-                    onClick={() => setShowModal(false)}
-                    className={classes.buttonReverted}
-                  >
-                    Close
-                  </button>
-                </div>
-              </div>
-            )}
+              </span>
+            </Link>
           </section>
         </main>
       )}
@@ -699,94 +586,12 @@ export default function GPTsSlug({ initialPageData }) {
               </Link>
             </div>
           </div>
-          <p className={classes.promoText}>
+          <Link href={portfolioLink} className={classes.promoText}>
             {translation.promo[activeLanguage]}{" "}
-            <Link
-              href={portfolioLink}
-              className={`${classes.portfolioLinkOrange}`}
-            >
-              {" "}
+            <span className={classes.portfolioLinkOrange}>
               {activeLanguage === "FR" ? "ici" : "here"}
-            </Link>{" "}
-          </p>
-          <div className={classes.donationBanner}>
-            <button
-              onClick={() => setShowModal(true)}
-              className={classes.clickHere}
-            >
-              {translation.donationBanner[activeLanguage]}
-            </button>
-          </div>
-          {showModal && (
-            <div className={classes.overlay}>
-              <div className={classes.modal}>
-                <h2 className={classes.titleDonation}>
-                  {translation.donationTitle[activeLanguage]}
-                </h2>
-                <p>{translation.donationText[activeLanguage]}</p>
-
-                <div className={classes.donationOptions}>
-                  <div className={classes.donationOption}>
-                    <p className={classes.donationText}>Paypal: </p>
-                    <div className={classes.donationCopy}>
-                      <p className={classes.donationInfos}>@wiseduckdev</p>
-                      <button onClick={() => handleCopy("@wiseduckdev")}>
-                        <BsCopy />
-                      </button>
-                    </div>
-                  </div>
-                  <div className={classes.donationOption}>
-                    <p className={classes.donationText}>
-                      {translation.BTC[activeLanguage]}
-                    </p>
-                    <div className={classes.donationCopy}>
-                      <p className={classes.donationInfos}>
-                        bc1qjalnp6rhvalxeyclcwfcp0madsmghcfzmylxf4
-                      </p>
-                      <button
-                        onClick={() =>
-                          handleCopy(
-                            "bc1qjalnp6rhvalxeyclcwfcp0madsmghcfzmylxf4"
-                          )
-                        }
-                      >
-                        <BsCopy />
-                      </button>
-                    </div>
-                  </div>
-                  <div className={classes.donationOption}>
-                    <p className={classes.donationText}>
-                      {translation.ETH[activeLanguage]}
-                    </p>
-                    <div className={classes.donationCopy}>
-                      <p className={classes.donationInfos}>
-                        0x97553A534aD05fb7D67aFa8fd4d8BCC8C990b477
-                      </p>
-                      <button
-                        onClick={() =>
-                          handleCopy(
-                            "0x97553A534aD05fb7D67aFa8fd4d8BCC8C990b477"
-                          )
-                        }
-                      >
-                        <BsCopy />
-                      </button>
-                    </div>
-                  </div>
-                  {copiedInfo && (
-                    <p className={classes.copiedNotification}>{copiedInfo}</p>
-                  )}
-                </div>
-
-                <button
-                  onClick={() => setShowModal(false)}
-                  className={classes.buttonReverted}
-                >
-                  Close
-                </button>
-              </div>
-            </div>
-          )}
+            </span>
+          </Link>
         </main>
       )}
